requirements.txt
keras==2.13.1
keras-facenet==0.3.2
mtcnn==0.1.1
numpy==1.24.3
opencv-python==4.9.0.80
Pillow==7.0.0
tensorflow==2.13.1
